LayoutTests/platform/chromium-mac/perf/mouse-event-expected.txt

Tests that mouse event dispatch is linear with respect to the number of parent elements.
FAIL: got indeterminate result expected O(n)

DEBUG LOG:
run iteration. magnitude 512 milliseconds 50 runsPerIteration 1
iterations 1
run iteration. magnitude 512 milliseconds 250 runsPerIteration 1
iterations 25
run iteration. magnitude 1 milliseconds 250 runsPerIteration 1
run iteration. magnitude 2 milliseconds 250 runsPerIteration 1
run iteration. magnitude 4 milliseconds 250 runsPerIteration 1
run iteration. magnitude 8 milliseconds 250 runsPerIteration 1
run iteration. magnitude 16 milliseconds 250 runsPerIteration 1
run iteration. magnitude 32 milliseconds 250 runsPerIteration 1
run iteration. magnitude 64 milliseconds 250 runsPerIteration 1
run iteration. magnitude 128 milliseconds 250 runsPerIteration 1
run iteration. magnitude 256 milliseconds 250 runsPerIteration 1
magnitude: 1 iterations: 2541 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.09838646202282567
magnitude: 2 iterations: 2638 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.09476876421531463
magnitude: 4 iterations: 2753 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.09081002542680712
magnitude: 8 iterations: 2172 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.1151012891344383
magnitude: 16 iterations: 1908 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.1310272536687631
magnitude: 32 iterations: 1364 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.18328445747800587
magnitude: 64 iterations: 816 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.30637254901960786
magnitude: 128 iterations: 388 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.6443298969072165
magnitude: 256 iterations: 142 runsPerIteration 1 loop-time 250 time/iteration(ms): 1.7605633802816902
magnitude: 512 iterations: 25 runsPerIteration 1 loop-time 250 time/iteration(ms): 10
magnitudes: 1,2,4,8,16,32,64,128,256,512
iterations: 2541,2638,2753,2172,1908,1364,816,388,142,25
milliseconds/iteration: 0.09838646202282567,0.09476876421531463,0.09081002542680712,0.1151012891344383,0.1310272536687631,0.18328445747800587,0.30637254901960786,0.6443298969072165,1.7605633802816902,10
numPoints 10 slope 0.017596854517936102 intercept -0.4576938093693963 rSquared 0.8849994862530944
numPoints 10 slope 0.9108105789436256 intercept -1.498501623869788 rSquared 0.38379086104585913
numPoints 10 slope 0.6567881167712596 intercept -3.2205728003077234 rSquared 0.7826211684740576
rSquared 0.8849994862530944 rSquaredXLog 0.38379086104585913 rSquaredXYLog 0.7826211684740576
numTriesLeft: 2
run iteration. magnitude 512 milliseconds 250 runsPerIteration 1
iterations 32
run iteration. magnitude 1 milliseconds 250 runsPerIteration 1
run iteration. magnitude 2 milliseconds 250 runsPerIteration 1
run iteration. magnitude 4 milliseconds 250 runsPerIteration 1
run iteration. magnitude 8 milliseconds 250 runsPerIteration 1
run iteration. magnitude 16 milliseconds 250 runsPerIteration 1
run iteration. magnitude 32 milliseconds 250 runsPerIteration 1
run iteration. magnitude 64 milliseconds 250 runsPerIteration 1
run iteration. magnitude 128 milliseconds 250 runsPerIteration 1
run iteration. magnitude 256 milliseconds 250 runsPerIteration 1
magnitude: 1 iterations: 2796 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.08941344778254649
magnitude: 2 iterations: 2806 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.08909479686386315
magnitude: 4 iterations: 2828 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.0884016973125884
magnitude: 8 iterations: 2577 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.09701202949165696
magnitude: 16 iterations: 1891 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.13220518244315177
magnitude: 32 iterations: 1775 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.14084507042253522
magnitude: 64 iterations: 901 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.27746947835738067
magnitude: 128 iterations: 355 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.704225352112676
magnitude: 256 iterations: 173 runsPerIteration 1 loop-time 250 time/iteration(ms): 1.4450867052023122
magnitude: 512 iterations: 32 runsPerIteration 1 loop-time 250 time/iteration(ms): 7.8125
magnitudes: 1,2,4,8,16,32,64,128,256,512
iterations: 2796,2806,2828,2577,1891,1775,901,355,173,32
milliseconds/iteration: 0.08941344778254649,0.08909479686386315,0.0884016973125884,0.09701202949165696,0.13220518244315177,0.14084507042253522,0.27746947835738067,0.704225352112676,1.4450867052023122,7.8125
numPoints 10 slope 0.013775436073809743 intercept -0.3216017343518658 rSquared 0.8946818919605138
numPoints 10 slope 0.7224741589220867 intercept -1.165888792280309 rSquared 0.39835304121075765
numPoints 10 slope 0.6411460735706459 intercept -3.2782011433027467 rSquared 0.7861379133323082
rSquared 0.8946818919605138 rSquaredXLog 0.39835304121075765 rSquaredXYLog 0.7861379133323082
numTriesLeft: 1
run iteration. magnitude 512 milliseconds 250 runsPerIteration 1
iterations 30
run iteration. magnitude 1 milliseconds 250 runsPerIteration 1
run iteration. magnitude 2 milliseconds 250 runsPerIteration 1
run iteration. magnitude 4 milliseconds 250 runsPerIteration 1
run iteration. magnitude 8 milliseconds 250 runsPerIteration 1
run iteration. magnitude 16 milliseconds 250 runsPerIteration 1
run iteration. magnitude 32 milliseconds 250 runsPerIteration 1
run iteration. magnitude 64 milliseconds 250 runsPerIteration 1
run iteration. magnitude 128 milliseconds 250 runsPerIteration 1
run iteration. magnitude 256 milliseconds 250 runsPerIteration 1
magnitude: 1 iterations: 2726 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.09170946441672781
magnitude: 2 iterations: 3794 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.06589351607801792
magnitude: 4 iterations: 2718 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.09197939661515821
magnitude: 8 iterations: 2313 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.10808473843493299
magnitude: 16 iterations: 2312 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.10813148788927336
magnitude: 32 iterations: 1147 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.21795989537925023
magnitude: 64 iterations: 980 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.25510204081632654
magnitude: 128 iterations: 362 runsPerIteration 1 loop-time 250 time/iteration(ms): 0.6906077348066298
magnitude: 256 iterations: 161 runsPerIteration 1 loop-time 250 time/iteration(ms): 1.5527950310559007
magnitude: 512 iterations: 30 runsPerIteration 1 loop-time 250 time/iteration(ms): 8.333333333333334
magnitudes: 1,2,4,8,16,32,64,128,256,512
iterations: 2726,3794,2718,2313,2312,1147,980,362,161,30
milliseconds/iteration: 0.09170946441672781,0.06589351607801792,0.09197939661515821,0.10808473843493299,0.10813148788927336,0.21795989537925023,0.25510204081632654,0.6906077348066298,1.5527950310559007,8.333333333333334
numPoints 10 slope 0.014705725526025603 intercept -0.3528360574298639 rSquared 0.8940855188799465
numPoints 10 slope 0.7705473334965658 intercept -1.251902538772523 rSquared 0.3973473837631844
numPoints 10 slope 0.6650403210160905 intercept -3.3387368774324435 rSquared 0.8069908755069581
rSquared 0.8940855188799465 rSquaredXLog 0.3973473837631844 rSquaredXYLog 0.8069908755069581
